+++ examples/monolog.php
@@ -3,6 +3,7 @@
 require_once __DIR__ . '/../vendor/autoload.php';
 
 use Elven\Observability\PhpLegacy\Logs\MonologTraceProcessor;
+use Elven\Observability\PhpLegacy\Logs\MonologOtlpHandler;
 use Elven\Observability\PhpLegacy\Observability;
 use Monolog\Handler\StreamHandler;
 use Monolog\Logger;
@@ -12,9 +13,10 @@ Observability::init(array('service_name' => 'monolog-example'));
 $logger = new Logger('example');
 $logger->pushProcessor(new MonologTraceProcessor());
 $logger->pushHandler(new StreamHandler('php://stdout'));
+$logger->pushHandler(new MonologOtlpHandler());
 
 Observability::tracer()->withSpan('log-example', function () use ($logger) {
-    $logger->info('safe event', array('operation' => 'demo'));
+    $logger->info('safe event', array('operation' => 'demo', 'token' => 'redacted-by-default'));
 });
 
 Observability::shutdown();